backend/requirements.txt
fastapi
uvicorn[standard]
motor
google-genai
python-dotenv
python-multipart
pypdf
web3